Update Drawer, DebugVoucherStepDrawer and DebugVoucherResults and their tests and snapshots
