+++ src/pages/ConsumerDebugVoucherPage/components/__test__/DebugVoucherResults.test.tsx
@@ -5,7 +5,6 @@ import { DebugVoucherResults } from '../DebugVoucherResults'
 import { fireEvent, render } from '@testing-library/react'
 import {
   createMockDebugVoucherRequest,
-  createMockDebugVoucherResultFailed,
   createMockDebugVoucherResultPassed,
   createMockDebugVoucherResultStep,
 } from '__mocks__/data/voucher.mocks'
@@ -61,249 +60,5 @@ describe('DebugVoucherResults testing', () => {
     fireEvent.click(newRequestButton)
 
     expect(handleMakeNewRequestMockFn).toBeCalled()
-  })
-
-  it('should setDebugVoucherStepDrawer function be called if close icon is clicked', () => {
-    const setDebugVoucherStepDrawerMockFn = vi.fn()
-
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultPassed(),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['clientAssertionValidation', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-      setDebugVoucherStepDrawer: setDebugVoucherStepDrawerMockFn,
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    const closeDrawerButton = screen.getByLabelText('closeIconAriaLabel')
-
-    fireEvent.click(closeDrawerButton)
-
-    expect(setDebugVoucherStepDrawerMockFn).toBeCalled()
-  })
-
-  it('should button nextStep not be rendered if clientKind is CONUSMER and selectedStep key is platformStatesVerification', () => {
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultPassed({ clientKind: 'CONSUMER' }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['platformStatesVerification', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    expect(screen.queryByRole('button', { name: 'stepDrawer.nextStepBtn' })).not.toBeInTheDocument()
-  })
-
-  it('should button nextStep not be rendered if clientKind is API and selectedStep key is clientAssertionSignatureVerification', () => {
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultPassed({
-        clientKind: 'API',
-        eservice: undefined,
-        steps: {
-          clientAssertionValidation: {
-            result: 'PASSED',
-            failures: [],
-          },
-          publicKeyRetrieve: {
-            result: 'PASSED',
-            failures: [],
-          },
-          clientAssertionSignatureVerification: {
-            result: 'PASSED',
-            failures: [],
-          },
-          platformStatesVerification: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-        },
-      }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['clientAssertionSignatureVerification', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    expect(screen.queryByRole('button', { name: 'stepDrawer.nextStepBtn' })).not.toBeInTheDocument()
-  })
-
-  it('should button nextStep not be rendered if clientKind is undefined and selectedStep key is clientAssertionSignatureVerification', () => {
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultFailed({
-        clientKind: undefined,
-        eservice: undefined,
-        steps: {
-          clientAssertionValidation: {
-            result: 'FAILED',
-            failures: [
-              {
-                code: 'test code',
-                reason: 'test reason',
-              },
-            ],
-          },
-          publicKeyRetrieve: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-          clientAssertionSignatureVerification: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-          platformStatesVerification: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-        },
-      }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['clientAssertionSignatureVerification', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    expect(
-      screen.getByText('stepDrawer.title.clientAssertionSignatureVerification')
-    ).toBeInTheDocument()
-
-    expect(screen.queryByRole('button', { name: 'stepDrawer.nextStepBtn' })).not.toBeInTheDocument()
-  })
-
-  it('should button nextStep be rendered if clientKind is CONUSMER and selectedStep key is not platformStatesVerification', () => {
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultPassed({ clientKind: 'CONSUMER' }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['clientAssertionValidation', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    expect(screen.getByText('stepDrawer.title.clientAssertionValidation')).toBeInTheDocument()
-
-    expect(screen.queryByRole('button', { name: 'stepDrawer.nextStepBtn' })).toBeInTheDocument()
-  })
-
-  it('should button nextStep be rendered if clientKind is API and selectedStep key is not clientAssertionSignatureVerification', () => {
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultPassed({
-        clientKind: 'API',
-        eservice: undefined,
-        steps: {
-          clientAssertionValidation: {
-            result: 'PASSED',
-            failures: [],
-          },
-          publicKeyRetrieve: {
-            result: 'PASSED',
-            failures: [],
-          },
-          clientAssertionSignatureVerification: {
-            result: 'PASSED',
-            failures: [],
-          },
-          platformStatesVerification: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-        },
-      }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['publicKeyRetrieve', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    expect(screen.getByText('stepDrawer.title.publicKeyRetrieve')).toBeInTheDocument()
-
-    expect(screen.getByRole('button', { name: 'stepDrawer.nextStepBtn' })).toBeInTheDocument()
-  })
-
-  it('should button nextStep be rendered if clientKind is undefined and selectedStep key is not clientAssertionSignatureVerification', () => {
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultFailed({
-        clientKind: undefined,
-        eservice: undefined,
-        steps: {
-          clientAssertionValidation: {
-            result: 'FAILED',
-            failures: [
-              {
-                code: 'test code',
-                reason: 'test reason',
-              },
-            ],
-          },
-          publicKeyRetrieve: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-          clientAssertionSignatureVerification: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-          platformStatesVerification: {
-            result: 'SKIPPED',
-            failures: [],
-          },
-        },
-      }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['clientAssertionValidation', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: vi.fn(),
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    expect(screen.getByText('stepDrawer.title.clientAssertionValidation')).toBeInTheDocument()
-
-    expect(screen.queryByRole('button', { name: 'stepDrawer.nextStepBtn' })).toBeInTheDocument()
-  })
-
-  it('should nextStep function be called on nextStepButton click', () => {
-    const goToNextStepMockFn = vi.fn()
-    mockDebugVoucherContext({
-      request: createMockDebugVoucherRequest(),
-      response: createMockDebugVoucherResultPassed({ clientKind: 'CONSUMER' }),
-      debugVoucherStepDrawer: {
-        isOpen: true,
-        selectedStep: ['clientAssertionValidation', createMockDebugVoucherResultStep()],
-      },
-      goToNextStep: goToNextStepMockFn,
-    })
-
-    const screen = render(<DebugVoucherResults />)
-
-    const nextStepButton = screen.getByRole('button', { name: 'stepDrawer.nextStepBtn' })
-
-    fireEvent.click(nextStepButton)
-
-    expect(goToNextStepMockFn).toBeCalled()
   })
 })
